refactor(core): extract IDatePickerProps to types.ts, use named ReactNode
Standalone interface with external type imports (TDateDisplayType,
TDayOfWeek, PopoverProps). No runtime change.
